chore: Reorganized the test utilities in AWS versioned tests
# Conflicts:
#	test/versioned/aws-sdk-v3/client-dynamodb.test.js
