+++ lib/transaction/trace/index.js
@@ -22,6 +22,8 @@ const SegmentTree = require('./segment-tree')
  * serialization of the transaction trace.
  *
  * @param {Transaction} transaction The transaction bound to the trace.
+ *
+ * @class {object} TransactionTrace
  */
 function Trace(transaction) {
   if (!transaction) {
