+++ test/versioned/aws-sdk-v2/http-services.test.js
@@ -4,12 +4,18 @@
  */
 
 'use strict'
+
 const assert = require('node:assert')
 const test = require('node:test')
+
 const helper = require('../../lib/agent_helper')
-const common = require('../aws-sdk-v3/common')
+const checkAWSAttributes = require('../aws-sdk-v3/test-utils/check-aws-attributes.js')
 const { createEmptyResponseServer, FAKE_CREDENTIALS } = require('../../lib/aws-server-stubs')
 const { match } = require('../../lib/custom-assertions')
+const {
+  EXTERN_PATTERN,
+  SEGMENT_DESTINATION
+} = require('../aws-sdk-v3/test-utils/constants.js')
 
 test('AWS HTTP Services', async (t) => {
   t.beforeEach(async (ctx) => {
@@ -238,13 +244,13 @@ test('AWS HTTP Services', async (t) => {
 })
 
 function finish(end, service, operation, tx) {
-  const externals = common.checkAWSAttributes({
+  const externals = checkAWSAttributes({
     trace: tx.trace,
     segment: tx.trace.root,
-    pattern: common.EXTERN_PATTERN
+    pattern: EXTERN_PATTERN
   })
   if (assert.equal(externals.length, 1, 'should have an aws external')) {
-    const attrs = externals[0].attributes.get(common.SEGMENT_DESTINATION)
+    const attrs = externals[0].attributes.get(SEGMENT_DESTINATION)
     match(attrs, {
       'aws.operation': operation,
       'aws.requestId': String,
