+++ test/versioned/aws-sdk-v3/rekognition.test.js
@@ -7,7 +7,8 @@
 
 const test = require('node:test')
 const helper = require('../../lib/agent_helper')
-const { afterEach, checkExternals } = require('./common')
+const afterEach = require('./test-utils/after-each.js')
+const checkExternals = require('./test-utils/check-externals.js')
 const { createEmptyResponseServer, FAKE_CREDENTIALS } = require('../../lib/aws-server-stubs')
 
 test('RekognitionClient', async (t) => {
