+++ test/versioned/aws-sdk-v3/sns.test.js
@@ -4,13 +4,21 @@
  */
 
 'use strict'
+
 const assert = require('node:assert')
 const test = require('node:test')
-const helper = require('../../lib/agent_helper')
-const common = require('./common')
-const awsEcho = require('./test-utils/aws-echo.js')
-const { createResponseServer, FAKE_CREDENTIALS } = require('../../lib/aws-server-stubs')
 const { tspl } = require('@matteo.collina/tspl')
+
+const helper = require('../../lib/agent_helper')
+const awsEcho = require('./test-utils/aws-echo.js')
+const checkAWSAttributes = require('./test-utils/check-aws-attributes.js')
+const checkExternals = require('./test-utils/check-externals.js')
+const {
+  EXTERN_PATTERN,
+  SEGMENT_DESTINATION,
+  SNS_PATTERN
+} = require('./test-utils/constants.js')
+const { createResponseServer, FAKE_CREDENTIALS } = require('../../lib/aws-server-stubs')
 const { match } = require('../../lib/custom-assertions')
 
 test('SNS', async (t) => {
@@ -132,7 +140,7 @@ test('SNS', async (t) => {
         await sns.send(cmd)
         tx.end()
 
-        setImmediate(common.checkExternals, {
+        setImmediate(checkExternals, {
           end,
           tx,
           service: 'SNS',
@@ -222,22 +230,22 @@ test('SNS', async (t) => {
 function finish(end, tx, destName) {
   const root = tx.trace.root
 
-  const messages = common.checkAWSAttributes({
+  const messages = checkAWSAttributes({
     trace: tx.trace,
     segment: root,
-    pattern: common.SNS_PATTERN
+    pattern: SNS_PATTERN
   })
   assert.equal(messages.length, 1, 'should have 1 message broker segment')
   assert.ok(messages[0].name.endsWith(destName), 'should have appropriate destination')
 
-  const externalSegments = common.checkAWSAttributes({
+  const externalSegments = checkAWSAttributes({
     trace: tx.trace,
     segment: root,
-    pattern: common.EXTERN_PATTERN
+    pattern: EXTERN_PATTERN
   })
   assert.equal(externalSegments.length, 0, 'should not have any External segments')
 
-  const attrs = messages[0].attributes.get(common.SEGMENT_DESTINATION)
+  const attrs = messages[0].attributes.get(SEGMENT_DESTINATION)
   match(attrs, {
     'aws.operation': 'PublishCommand',
     'aws.requestId': String,
