+++ test/versioned/aws-sdk-v3/sqs.test.js
@@ -9,8 +9,14 @@ const assert = require('node:assert')
 const test = require('node:test')
 
 const helper = require('../../lib/agent_helper')
-const common = require('./common')
 const awsEcho = require('./test-utils/aws-echo.js')
+const checkAWSAttributes = require('./test-utils/check-aws-attributes.js')
+const afterEach = require('./test-utils/after-each.js')
+const {
+  EXTERN_PATTERN,
+  SEGMENT_DESTINATION,
+  SQS_PATTERN
+} = require('./test-utils/constants.js')
 const { createResponseServer, FAKE_CREDENTIALS } = require('../../lib/aws-server-stubs')
 const { match } = require('../../lib/custom-assertions')
 
@@ -45,7 +51,7 @@ test('SQS API', async (t) => {
   })
 
   t.afterEach((ctx) => {
-    common.afterEach(ctx)
+    afterEach(ctx)
   })
 
   await t.test('commands with promises', async (t) => {
@@ -179,10 +185,10 @@ function finish({ transaction, queueName }) {
   const expectedSegmentCount = 3
 
   const root = transaction.trace.root
-  const segments = common.checkAWSAttributes({
+  const segments = checkAWSAttributes({
     trace: transaction.trace,
     segment: root,
-    pattern: common.SQS_PATTERN
+    pattern: SQS_PATTERN
   })
 
   assert.equal(
@@ -191,10 +197,10 @@ function finish({ transaction, queueName }) {
     `should have ${expectedSegmentCount} AWS MessageBroker/SQS segments`
   )
 
-  const externalSegments = common.checkAWSAttributes({
+  const externalSegments = checkAWSAttributes({
     trace: transaction.trace,
     segment: root,
-    pattern: common.EXTERN_PATTERN
+    pattern: EXTERN_PATTERN
   })
   assert.equal(externalSegments.length, 0, 'should not have any External segments')
 
@@ -226,7 +232,7 @@ function checkName(name, action, queueName) {
 }
 
 function checkAttributes(segment, operation) {
-  const actualAttributes = segment.attributes.get(common.SEGMENT_DESTINATION)
+  const actualAttributes = segment.attributes.get(SEGMENT_DESTINATION)
 
   const expectedAttributes = {
     'aws.operation': operation,
